Global sub id v2plus (#9884)
* implement migrate method in coordinator v2 plus

* add extra args field to requestRandomWords()

* add missing client file

* prettier

* prettier

* add forge tests for migrations

* address comments

* add missing file

* remove s_requestPayment()

* fix failing tests

* make subId uint256 instead of uint64. gas optimization for s_consumers

* run prettier

* minor changes

* remove vrf v2 plus subscription manager

* address comments

* regen wrappers

+++ contracts/src/v0.8/dev/interfaces/IVRFCoordinatorV2Plus.sol
@@ -52,7 +52,7 @@ interface IVRFCoordinatorV2Plus {
    * @dev    amount,
    * @dev    abi.encode(subId));
    */
-  function createSubscription() external returns (uint64 subId);
+  function createSubscription() external returns (uint256 subId);
 
   /**
    * @notice Get a VRF subscription.
@@ -63,7 +63,7 @@ interface IVRFCoordinatorV2Plus {
    * @return consumers - list of consumer address which are able to use this subscription.
    */
   function getSubscription(
-    uint64 subId
+    uint256 subId
   ) external view returns (uint96 balance, uint96 ethBalance, address owner, address[] memory consumers);
 
   /**
@@ -71,7 +71,7 @@ interface IVRFCoordinatorV2Plus {
    * @param subId - ID of the subscription
    * @param newOwner - proposed new owner of the subscription
    */
-  function requestSubscriptionOwnerTransfer(uint64 subId, address newOwner) external;
+  function requestSubscriptionOwnerTransfer(uint256 subId, address newOwner) external;
 
   /**
    * @notice Request subscription owner transfer.
@@ -79,28 +79,28 @@ interface IVRFCoordinatorV2Plus {
    * @dev will revert if original owner of subId has
    * not requested that msg.sender become the new owner.
    */
-  function acceptSubscriptionOwnerTransfer(uint64 subId) external;
+  function acceptSubscriptionOwnerTransfer(uint256 subId) external;
 
   /**
    * @notice Add a consumer to a VRF subscription.
    * @param subId - ID of the subscription
    * @param consumer - New consumer which can use the subscription
    */
-  function addConsumer(uint64 subId, address consumer) external;
+  function addConsumer(uint256 subId, address consumer) external;
 
   /**
    * @notice Remove a consumer from a VRF subscription.
    * @param subId - ID of the subscription
    * @param consumer - Consumer to remove from the subscription
    */
-  function removeConsumer(uint64 subId, address consumer) external;
+  function removeConsumer(uint256 subId, address consumer) external;
 
   /**
    * @notice Cancel a subscription
    * @param subId - ID of the subscription
    * @param to - Where to send the remaining LINK to
    */
-  function cancelSubscription(uint64 subId, address to) external;
+  function cancelSubscription(uint256 subId, address to) external;
 
   /*
    * @notice Check to see if there exists a request commitment consumers
@@ -109,5 +109,5 @@ interface IVRFCoordinatorV2Plus {
    * @return true if there exists at least one unfulfilled request for the subscription, false
    * otherwise.
    */
-  function pendingRequestExists(uint64 subId) external view returns (bool);
+  function pendingRequestExists(uint256 subId) external view returns (bool);
 }
